style single ping page
