+++ src/components/SinglePing/Comment.js
@@ -1,8 +1,7 @@
 import React from "react";
-import { Grid, Typography, Avatar, Paper } from "@material-ui/core";
+import { Avatar } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import { FaUser } from "react-icons/fa";
-import { Link } from "react-router-dom";
 import moment from "moment";
 
 import StyledComment from "../Styled/StyledComment";
